Skift til labymod events
