+++ core/src/main/java/ml/volder/transporter/modules/mcmmomodule/McmmoManager.java
@@ -4,7 +4,6 @@ import ml.volder.transporter.modules.mcmmomodule.skills.AbilitySkill;
 import ml.volder.transporter.modules.mcmmomodule.skills.SimpleSkill;
 import ml.volder.unikapi.datasystem.Data;
 import ml.volder.unikapi.datasystem.DataManager;
-import ml.volder.unikapi.event.events.clientmessageevent.ClientMessageEvent;
 import ml.volder.unikapi.types.Material;
 import ml.volder.unikapi.widgets.ModuleSystem;
