+++ core/src/main/java/ml/volder/transporter/modules/mcmmomodule/skills/AbilitySkill.java
@@ -2,7 +2,6 @@ package ml.volder.transporter.modules.mcmmomodule.skills;
 
 import ml.volder.unikapi.datasystem.Data;
 import ml.volder.unikapi.datasystem.DataManager;
-import ml.volder.unikapi.event.events.clientmessageevent.ClientMessageEvent;
 import ml.volder.unikapi.types.ModColor;
 import ml.volder.unikapi.widgets.ModuleSystem;
