+++ core/src/main/java/ml/volder/unikapi/api/ApiReferenceStorage.java
@@ -17,15 +17,8 @@ public interface ApiReferenceStorage {
     MinecraftAPI getMinecraftAPI();
     PlayerAPI getPlayerAPI();
 
-    Class<? extends EventImpl> getVersionedClientKeyPressEvent();
-    Class<? extends EventImpl> getVersionedClientMessageEvent();
-    Class<? extends EventImpl> getVersionedClientTickEvent();
-    Class<? extends EventImpl> getVersionedDrawScreenEvent();
     Class<? extends EventImpl> getVersionedOpenSignEvent();
     Class<? extends EventImpl> getVersionedMainMenuOpenEvent();
-    Class<? extends EventImpl> getVersionedSendMessageEvent();
-    Class<? extends EventImpl> getVersionedServerSwitchEvent();
-
 
     KeyMapper getVersionedKeyMapper();
