+++ core/src/main/java/ml/volder/unikapi/widgets/GuiModuleRenderer.java
@@ -3,9 +3,7 @@ package ml.volder.unikapi.widgets;
 
 import ml.volder.unikapi.datasystem.Data;
 import ml.volder.unikapi.datasystem.DataManager;
-import ml.volder.unikapi.event.EventHandler;
 import ml.volder.unikapi.event.Listener;
-import ml.volder.unikapi.event.events.drawscreenevent.DrawScreenEvent;
 
 public class GuiModuleRenderer implements Listener {
 
@@ -15,13 +13,10 @@ public class GuiModuleRenderer implements Listener {
         this.moduleManager = moduleManager;
     }
 
-    @EventHandler
-    public void drawScreenEvent(DrawScreenEvent event) {
+    public void drawScreenEvent(/*DrawScreenEvent event*/) {
         if(!ModuleSystem.shouldRenderModules())
             return;
-        moduleManager.getGuiModuleList().forEach(guiModule -> {
-            if(guiModule.isEnabled())
-                guiModule.drawModule();
-        });
+        // Not used anymore - drawn through the labymod widget system
+        //TODO remove old module system
     }
 }
